tie bridge rendering to their respective blocks
