+++ src/util/graphics/index.ts
@@ -3,3 +3,4 @@ export * from './block_asset'
 export * from './translate_pos'
 export * from './outline_image'
 export * from './draw_rotated'
+export * from './draw_bridge'
